requirements.txt
openai
python-dotenv
gspread
google-auth
requests
python-docx
streamlit
